+++ internal/configs/tomlConfigParser.go
@@ -23,7 +23,7 @@ import (
 
 	"github.com/BurntSushi/toml"
 	logger "github.com/sirupsen/logrus"
-	config "github.com/wso2/micro-gw/internal/configs/confTypes"
+	config "github.com/wso2/micro-gw/configs/confTypes"
 )
 
 var (
@@ -45,11 +45,11 @@ func ReadConfigs() (*config.Config, error) {
 		configs = new(config.Config)
 		mgwHome, _ := os.Getwd()
 		// logger.Info("MGW_HOME: ", mgwHome)
-		_, err := os.Stat(mgwHome + "/resources/conf/config.toml")
+		_, err := os.Stat(mgwHome + "/../resources/conf/config.toml")
 		if err != nil {
 			logger.Fatal("Configuration file not found.", err)
 		}
-		content, readErr := ioutil.ReadFile(mgwHome + "/resources/conf/config.toml")
+		content, readErr := ioutil.ReadFile(mgwHome + "/../resources/conf/config.toml")
 		if readErr != nil {
 			logger.Fatal("Error reading configurations. ", readErr)
 		}
@@ -69,11 +69,12 @@ func ReadLogConfigs() (*config.LogConfig, error) {
 	once_lc.Do(func() {
 		logConfigs = new(config.LogConfig)
 		mgwHome, _ := os.Getwd()
-		_, err := os.Stat(mgwHome + "/resources/conf/log_config.toml")
+		//TODO: (VirajSalaka) Provide path properly
+		_, err := os.Stat(mgwHome + "/../resources/conf/log_config.toml")
 		if err != nil {
 			logger.Fatal("Log configuration file not found.", err)
 		}
-		content, readErr := ioutil.ReadFile(mgwHome + "/resources/conf/log_config.toml")
+		content, readErr := ioutil.ReadFile(mgwHome + "/../resources/conf/log_config.toml")
 		if readErr != nil {
 			logger.Fatal("Error reading log configurations. ", readErr)
 		}
