+++ internal/pkg/api/restserver/operations/api_individual/post_import_api.go
@@ -25,7 +25,7 @@ import (
 
 	"github.com/go-openapi/runtime/middleware"
 
-	"github.com/wso2/micro-gw/internal/pkg/api/models"
+	"github.com/wso2/micro-gw/pkg/api/models"
 )
 
 // PostImportAPIHandlerFunc turns a function with the right signature into a post import API handler
