+++ internal/pkg/logging/rootLogs.go
@@ -23,7 +23,7 @@ import (
 	"strings"
 
 	"github.com/sirupsen/logrus"
-	"github.com/wso2/micro-gw/internal/configs"
+	"github.com/wso2/micro-gw/configs"
 	lumberjack "gopkg.in/natefinch/lumberjack.v2"
 )
