+++ internal/pkg/oasparser/envoyCodegen/httpFilters.go
@@ -24,7 +24,7 @@ import (
 	"github.com/envoyproxy/go-control-plane/pkg/wellknown"
 
 	"github.com/golang/protobuf/ptypes"
-	logger "github.com/wso2/micro-gw/internal/loggers"
+	logger "github.com/wso2/micro-gw/loggers"
 )
 
 /**
