+++ internal/pkg/oasparser/envoyCodegen/listener.go
@@ -26,8 +26,8 @@ import (
 	"github.com/envoyproxy/go-control-plane/pkg/wellknown"
 
 	"github.com/golang/protobuf/ptypes"
-	"github.com/wso2/micro-gw/internal/configs"
-	logger "github.com/wso2/micro-gw/internal/loggers"
+	"github.com/wso2/micro-gw/configs"
+	logger "github.com/wso2/micro-gw/loggers"
 )
 
 /**
