+++ internal/pkg/oasparser/envoyCodegen/routesWithClusters.go
@@ -25,10 +25,10 @@ import (
 	routev3 "github.com/envoyproxy/go-control-plane/envoy/config/route/v3"
 	envoy_type_matcherv3 "github.com/envoyproxy/go-control-plane/envoy/type/matcher/v3"
 
-	"github.com/wso2/micro-gw/internal/configs"
-	logger "github.com/wso2/micro-gw/internal/loggers"
-	"github.com/wso2/micro-gw/internal/pkg/oasparser/models/apiDefinition"
-	swag_operator "github.com/wso2/micro-gw/internal/pkg/oasparser/swaggerOperator"
+	"github.com/wso2/micro-gw/configs"
+	logger "github.com/wso2/micro-gw/loggers"
+	"github.com/wso2/micro-gw/pkg/oasparser/models/apiDefinition"
+	swag_operator "github.com/wso2/micro-gw/pkg/oasparser/swaggerOperator"
 
 	"strings"
 	"time"
