+++ internal/resources/adminAPI.yaml
