+++ internal/resources/license.txt
